Protecting routes with middleware

+++ routes/web.php
@@ -8,34 +8,34 @@ use App\Http\Controllers\UserController;
 Route::get('/', [ListingController::class, 'index']);
 
 // Show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Edit listing
-Route::get('/listings/{id}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{id}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Update listing
-Route::put('/listings/{id}', [ListingController::class, 'update']);
+Route::put('/listings/{id}', [ListingController::class, 'update'])->middleware('auth');
 
 // Store listing data
-Route::post('/listings/', [ListingController::class, 'store']);
+Route::post('/listings/', [ListingController::class, 'store'])->middleware('auth');
 
 // Delete listing
-Route::delete('/listings/{id}', [ListingController::class, 'delete']);
+Route::delete('/listings/{id}', [ListingController::class, 'delete'])->middleware('auth');
 
 // Single listing
 Route::get('/listings/{id}', [ListingController::class, 'show']);
 
 // Show Register form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // Create user
 Route::post('/users', [UserController::class, 'store']);
 
 // Logout user
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 // Login user
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Log in user
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
